move randomness tests into a single file and share the getrandom implementation across unices
